move regex functions in d to regex.c and regex.d
