+++ M2/Macaulay2/d/Makefile.files.in
@@ -10,7 +10,7 @@ ifeq "@DEBUG@" "yes"
 M2_OBJECTS += debug.o
 endif
 M2_CFILES += gmp_aux.c 
-M2_LASTCFILES += M2types.c scclib.c M2lib.c M2mem.c gdbm_interface.c
+M2_LASTCFILES += M2types.c scclib.c regex.c M2lib.c M2mem.c gdbm_interface.c
 M2_MADECFILES := startup.c
 M2_OBJECTS += $(M2_MADECFILES:.c=.o)
 M2_OBJECTS += ../regex/regex.o
@@ -58,6 +58,7 @@ M2_DFILES += basic.d
 M2_DFILES += convertr.d
 M2_DFILES += common.d
 M2_DFILES += util.d
+M2_DFILES += regex.d
 M2_DFILES += struct.d
 M2_DFILES += classes.dd
 M2_DFILES += buckets.dd
