+++ M2/Macaulay2/d/actors4.d
@@ -5,6 +5,7 @@ use actors;
 use actors2;
 use struct;
 use pthread;
+use regex;
 
 header "#include <engine.h>";
