+++ M2/Macaulay2/d/system.d
@@ -175,11 +175,6 @@ import isRegularFile(name:string):int;
 import wordexp(word:string):ArrayStringOrNull;
 import readlink(filename:string):string;
 import realpath(filename:string):(null or string);
-import noErrorMessage:string;
-import regexmatchErrorMessage:string;
-import regexmatch(pattern:string, start:int, range:int, text:string, ignorecase:bool):array(int);
-import regexreplace(pattern:string, replacement: string, text:string, errflag:string, ignorecase:bool):string;
-import regexselect(pattern:string, replacement: string, text:string, errflag:array(string), ignorecase:bool):array(string);
 import readDirectory(name:string):(null or array(string));
 import strncmp(s:string,t:string,n:int):int;
 import history():array(string);
